convertAdvancedQueryTo/FromString moved to own functions
