+++ src/config.js
@@ -2,5 +2,6 @@ export const DATABASE_URL = 'https://subject-set-search-api.zooniverse.org/'
 export const DATABASE_NAME = 'projects'
 export const TABLE_PREFIX = 'proj_'
 export const SUBJECT_ID_KEY = 'subject_id'
+export const KEYWORDS_KEY = '#keywords'
 export const PAGE_SIZE = 20
 export const ADVANCED_QUERY_PREFIX = 'advquery_'
